+++ tests/instrument.rs
@@ -17,7 +17,7 @@ const CLASSICAL_FSHARP4_48000: &str = include_str!("./samples/F-4_48000_classica
 #[case::D4_44100_Piano(PIANO_D4_44100, 293.7, 44100)]
 #[case::E2_44100_Acoustic(ACOUSTIC_E2_44100, 82.41, 44100)]
 #[case::FSharp4_48000_Classical(CLASSICAL_FSHARP4_48000, 370.0, 48000)]
-fn test_yin_instrument(#[case] data: &str, #[case] expected_freq: f64, #[case] sample_rate: usize) {
+fn test_yin_instrument(#[case] data: &str, #[case] expected_freq: f32, #[case] sample_rate: usize) {
     let pitch = run_yin(data, sample_rate, 0.15, None);
 
     approx::assert_relative_eq!(pitch, expected_freq, epsilon = 0.01 * expected_freq);
@@ -28,7 +28,7 @@ fn test_yin_instrument(#[case] data: &str, #[case] expected_freq: f64, #[case] s
 #[case::E2_44100_Acoustic(ACOUSTIC_E2_44100, 82.41, 44100, 75.0, 90.0)]
 fn test_yin_instrument_with_range(
     #[case] data: &str,
-    #[case] expected_freq: f64,
+    #[case] expected_freq: f32,
     #[case] sample_rate: usize,
     #[case] min: f64,
     #[case] max: f64,
@@ -46,7 +46,7 @@ fn test_yin_instrument_with_range(
 //#[case::FSharp4_48000_Classical(CLASSICAL_FSHARP4_48000, 370.0, 48000)] // fails - 186Hz detected?
 fn test_pyin_instrument(
     #[case] data: &str,
-    #[case] expected_freq: f64,
+    #[case] expected_freq: f32,
     #[case] sample_rate: usize,
 ) {
     let pitch = run_pyin(data, sample_rate, None);
@@ -59,7 +59,7 @@ fn test_pyin_instrument(
 #[case::FSharp4_48000_Classical(CLASSICAL_FSHARP4_48000, 370.0, 48000, 300.0, 400.0)]
 fn test_pyin_instrument_with_range(
     #[case] data: &str,
-    #[case] expected_freq: f64,
+    #[case] expected_freq: f32,
     #[case] sample_rate: usize,
     #[case] min: f64,
     #[case] max: f64,
